requirements.txt
asgiref==3.5.2
Babel==2.10.3
beautifulsoup4==4.11.1
Django==4.0.6
django-livereload-server==0.4
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
phonenumbers==8.12.52
Pillow==9.2.0
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
